Exclude common/int128.h from cpluspluscheck
It uses static assertions, which are not supported under C++ in this
branch.

This change only goes into the 9.4 branch, because 9.5 and beyond will
primarily use the USE_NATIVE_INT128 branch, so cpluspluscheck isn't
bothered.  In PG11 we will have C++ support for static assertions, so
the issue will go away altogether.