Fix build failure on MSVC.
Commit 824cceded introduced use of pg_malloc and pg_realloc into
specscanner.l, but it isn't working in 9.3 on MSVC.  Evidently we
added the infrastructure for that in 9.4.  Since the chance of an
actual OOM here is tiny, and the consequences would only be an
isolation test failure, and we have unchecked OOM hazards elsewhere
in this file in 9.3, it's not worth sweating over.  Just replace
the calls with malloc and realloc.

Per buildfarm.